Implement and (primitive) test Skein1024 optimization
Not yet used in brute-forcer.

In general it is not valid to process a single full Skein block immediately
during Update().  This is because, if it turns out to be the final input
block, it is processed differently.

For this application, however, we can cheat a little and provide advanced
knowledge that our first full block will not be the final input block.

Test that assumption with a single static example via '-S' option.